% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tdsp_structure.R
\name{create_tdsp_structure}
\alias{create_tdsp_structure}
\title{Copy the Azure TDSP project structure.}
\usage{
create_tdsp_structure(path = ".")
}
\arguments{
\item{path}{The path where to copy the TDSP template.}
}
\description{
Copies the \href{https://github.com/Azure/Azure-TDSP-ProjectTemplate}{TDSP template} to a new folder.
}
\examples{
where <- file.path(tempdir(), "TDSP")
create_tdsp_structure(where)
}
